Have getStats and getStatsId return full list of items from db

+++ src/main/java/com/francislainy/gatling_tool/dto/stats/MeanNumberOfRequestsPerSecond.java
@@ -15,6 +15,6 @@ public class MeanNumberOfRequestsPerSecond {
     public Double ok;
     @SerializedName("ko")
     @Expose
-    public Integer ko;
+    public Double ko;
 
 }
